Working Refresh, Refresh Button, and Preferences Button
In this commit, several things were added. In the previous commit, tapping inside the void space below the widget window was the only way to access the preferences. This commit introduces a Preferences button, which behaves as expected and improves usability. Additionally, a new Refresh button has been added below the widget window and it causes an instant refresh of the stocks without having to go into the preferences and clicking update.
This references issues #12 and #11.