Implemented Actual Columns, not just Formatted String for Rows
In this commit, introduced new file bonobo_row_item.xml, which is a layout that represents a single item in the row separated into three distinct TextViews. This will allow for more flexibility in the future in terms of coloring fields, access/mutation, etc. References Issue #20 since it will work well in conjunction with Graphs.